// frontend/src/api/generated/model/employeeCertificationDto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */
import type { EmployeeAttachmentDto } from './employeeAttachmentDto';

export interface EmployeeCertificationDto {
  id: string;
  name: string;
  issuedBy: string;
  issuedDate: string;
  /** @nullable */
  expiredDate: string | null;
  /** @nullable */
  attachment?: EmployeeAttachmentDto | null;
}